refactor: Move 'Admin' specific components into 'Admin' at 'Pages'
